requirements.txt
customtkinter>=5.2.0
PyMuPDF>=1.23.0
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
matplotlib>=3.7.0
Pillow>=10.0.0
openpyxl>=3.1.0
ultralytics>=8.0.0
pyinstaller>=6.0.0
